Update project structure, add graphql support and remove openapi_to_proto complex (#11)

+++ migrations/20250601190210_create_order_table_and_link_user.sql
@@ -3,7 +3,7 @@
 create table orders (
   id serial primary key,
   user_id int not null,
-  amount int not null,
+  amount float not null,
   created_at timestamptz default now(),
   updated_at timestamptz default now()
 );
